Change '\TemplateName' to '\AcroTemplateName'
There appears to be an error where \AcroTemplateName is incorrectly called \TemplateName in two places. If this is intentional please let me know. I modified these lines and it solved the issues I had with using the toc list format ((here is someone else with the same error)[https://tex.stackexchange.com/questions/561791/acro-package-list-same-style-as-table-of-contents]).